% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioseq.R
\name{check_completeness}
\alias{check_completeness}
\title{Transcriptome completeness checking}
\usage{
check_completeness(prot_seqs, prot_blast)
}
\arguments{
\item{prot_seqs}{Transcriptome sequences translated to proteins}

\item{prot_blast}{Proteins BLAST output}
}
\value{
A dataset containing the completeness calssification and coverage
}
\description{
\code{check_completeness()}verify if the transcripts were assembled completely and the
coverage of those sequences
}
