% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{check_taxonomy}
\alias{check_taxonomy}
\title{Check if the homologous proteins identified by local alignment come from a certain phylum}
\usage{
check_taxonomy(
  nr_matches,
  nr_taxonomy = NULL,
  tr_matches,
  tr_taxonomy = NULL,
  phylum_name = "Mollusca",
  evalmax = 1e-06
)
}
\arguments{
\item{nr_matches}{The BLASTP result against the NCBI non-redundant protein database (BLAST output format 6)}

\item{nr_taxonomy}{Dataset with information to convert NCBI accession number to taxonomy information}

\item{tr_matches}{The BLASTP result against the UniPrto/TrEMBL database (BLAST output format 6)}

\item{tr_taxonomy}{Dataset with information to convert UniProt accession number to taxonomy information}

\item{phylum_name}{Phylum name to check}

\item{evalmax}{Maximum e-value allowed to qualify a match}
}
\value{
If the matches correspond to the phylum specified
}
\description{
Check if the homologous proteins identified by local alignment come from a certain phylum
}
