% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_annotation.R
\name{go_annotation}
\alias{go_annotation}
\title{Transcripts functional annotation}
\usage{
go_annotation(trans_blast, trans_hmm, pfam_goid = NULL, unip_goid = NULL)
}
\arguments{
\item{trans_blast}{Result of BLAST alignment if should be output format 6.}

\item{trans_hmm}{Result of protein domain identification}

\item{pfam_goid}{Conversion table between protein domains and GO IDs}

\item{unip_goid}{Conversion table between protein accession in UniProt DB and GO IDS}
}
\value{
Conversion table between the transcripts name and GO IDs
}
\description{
\code{go_annotation()} Uses the information about homologous proteins and protein domains
detected in the transcripts to predict the Gene Ontologies associated with the sequences
assembled
}
