% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioseq.R
\name{rev_translate}
\alias{rev_translate}
\title{Reverse translate}
\usage{
rev_translate(aaseq)
}
\arguments{
\item{aaseq}{Protein sequence}
}
\value{
A DNA sequence.
}
\description{
\code{rev_translate()} accepts a protein sequence as input and uses the standard genetic code
to generate a DNA sequence representing the most likely non-degenerate coding sequence.
}
\examples{
sq <- "MACDEFGHIKL*"
rev_translate(sq)
}
